requirements.txt
pyyaml
django
pandas
beautifulsoup4
git+https://github.com/clombec/OrthoAParse.git@main#egg=orthoaget
